Rename authentication to auth
