+++ deployment/dev/compose/core.compose.yml
@@ -153,7 +153,7 @@ services:
       - replayer
 
   auth:
-    image: bazel/osiris/authentication:image
+    image: bazel/osiris/auth:image
     ports:
       - ${CEPTA_AUTH_BACKEND_PORT}:80
     volumes:
