+++ osiris/auth/test-auth.sh
