+++ osiris/usermgmt/BUILD
@@ -56,9 +56,9 @@ go_library(
         "//models/grpc:usermgmt_go_proto",
         "//models/types:result_go_proto",
         "//models/types:users_go_proto",
-        "//models/grpc:authentication_go_proto",
+        "//models/grpc:auth_go_proto",
         "//osiris/lib/db:go_default_library",
-        "//osiris/authentication:go_default_library",
+        "//osiris/auth:go_default_library",
         "//osiris/usermgmt/lib:go_default_library",
         "//osiris/lib:cli",
         "//osiris/lib:utils",
